--- app/src/main/java/com/example/cookingidea/ui/screen/main/MainUiState.kt
@@ -1,9 +1,0 @@
-package com.example.cookingidea.ui.screen.main
-
-import android.net.Uri
-
-data class MainUiState(
-    val selectedImageUri: Uri? = null,
-    val isLoading: Boolean = false,
-    val showDialog: Boolean = false,
-)
