--- app/src/main/java/com/example/cookingidea/ui/screen/menudialog/MenuDialogContent.kt
@@ -1,53 +1,0 @@
-package com.example.cookingidea.ui.screen.menudialog
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.example.cookingidea.R
-
-@Composable
-fun MenuDialogContent(
-    uiState: MenuDialogUiState,
-    modifier: Modifier = Modifier,
-) {
-
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(color = MaterialTheme.colorScheme.background)
-            .padding(20.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        Text(
-            text = stringResource(id = R.string.menu_dialog_title),
-            style = MaterialTheme.typography.headlineMedium,
-        )
-
-        Text(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .background(color = MaterialTheme.colorScheme.surfaceVariant)
-                .verticalScroll(rememberScrollState()),
-            text = uiState.generatedText,
-            style = MaterialTheme.typography.bodyMedium,
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun PreviewMenuDialogContent() {
-    MenuDialogContent(uiState = MenuDialogUiState())
-}
