--- app/src/main/java/com/example/cookingidea/ui/screen/menudialog/MenuDialogUiState.kt
@@ -1,5 +1,0 @@
-package com.example.cookingidea.ui.screen.menudialog
-
-data class MenuDialogUiState(
-    val generatedText: String = "",
-)
