+++ app/src/main/java/com/khaki/cookingidea/ui/screen/main/MainUiState.kt
@@ -1,0 +1,9 @@
+package com.khaki.cookingidea.ui.screen.main
+
+import android.net.Uri
+
+data class MainUiState(
+    val selectedImageUri: Uri? = null,
+    val isLoading: Boolean = false,
+    val showDialog: Boolean = false,
+)
