+++ app/src/main/java/com/khaki/cookingidea/ui/screen/menudialog/MenuDialogUiState.kt
@@ -1,0 +1,5 @@
+package com.khaki.cookingidea.ui.screen.menudialog
+
+data class MenuDialogUiState(
+    val generatedText: String = "",
+)
